Lookup for environmental data

Lookup for environmental data

env

Format

A data frame with 65 rows and 17 variables:

group

Broad type of environmental layer.

category

Another broad grouping of environmental layer.

provider

Who made this environmental layer?

data_name

The level at which a reference applies.

process

Three letter code identifying the process used to generate the layer

layer

Index of band/layer in process.

method

What method was used to summarise process.

static

Does the concept in this layer remain the same through time?

season_type

One of; year: whole year, each: split by season or NA

year_min

Minimum year available for data_name and prod.

transform

Numeric. Divide by transform to convert to units from numbers in layer.

units

Units for number in layer / transform.

description

Description of the data in prod.

detail

Other relevant details.

indicates

Why is the data in layer useful in predicting ecosystem distribution?

link

Link to webpage for data_name.

link_md

link formatted for use in rmarkdown.

...