Make a description for an ecosystem

Make a description for an ecosystem

make_eco_desc(
  bio_df,
  clust_df,
  bio_wide = NULL,
  bio_ind = NULL,
  context,
  clust_col = "cluster",
  cov_col = "use_cover",
  taxa_col = "taxa",
  str_col = "lifeform",
  ind_col = "ind",
  lustr,
  taxonomy,
  use_prop_thresh,
  use_p_val = 0.05
)

Arguments

bio_df

Dataframe with biological information.

clust_df

Dataframe with cluster membership and join columns to bio_df.

bio_wide

Dataframe of taxa by sites.

bio_ind

Dataframe of indicator taxa per class. Output from envCluster::make_ind_val_df. Will be created if not supplied.

context

Character name of columns in bio_df that define the context.

clust_col

Character name of column in clust_df that defines clusters.

cov_col

Character name of column in bio_df that contain numeric cover values.

taxa_col

Character name of column with taxa.

lustr

Dataframe containing structural information.

taxonomy

Dataframe containing indigenous status of taxa in bio_df

use_prop_thresh

Numeric. Threshold (proportion) for taxa to include in description. Taxa that occur in more than use_prop_thresh proportion of sites in the cluster will be included in the description.

use_p_val

Numeric 0 to 1. The p-value to use to accept a taxa as an indicator for an ecosystem.