Make a description for an ecosystem

Make a description for an ecosystem

make_eco_desc(
  bio_df,
  bio_wide,
  clust_df,
  ind_val_df = NULL,
  context,
  clust_col = "cluster",
  taxa_col = "taxa",
  cov_col = "cover_adj",
  ind_abu_col = "use_cover",
  str_col = "lifeform",
  lustr,
  taxonomy,
  use_prop_thresh = 0.8,
  use_p_val = 0.05,
  n_ind_max = 3,
  ind_val_iter = 3000,
  wt_ht_quant = 0.5,
  sites_sf_quant = 0.75,
  ht_sf_quant = 0.5,
  sites_sf_taxa_quant = 0.95
)

Arguments

bio_df

Dataframe containing the site and taxa data in long format.

bio_wide

Wide version of bio_df.

clust_df

Dataframe with context column(s) and a column with cluster membership for that context. Optional if clust_col appears in bio_df.

ind_val_df

Result from envCluster::make_ind_val_df(). If not provided it will be made from the other provided dfs (which is slow).

context

Character. Name(s) of column(s) that define the context.

clust_col

Character. Name of column containing cluster membership.

taxa_col

Character. Name of column containing the taxa names.

cov_col

Character. Name of column containing numeric abundance data ( usually 'cover' for plants).

ind_abu_col

Character. Name of column containing numeric abundance data for use in indicator analysis (labdsv::indval()). This can be the same as cov_col but at times a different measure of abundance is helpful here.

str_col

Character name of column in bio_df containing lifeform (or structural) information.

lustr

Dataframe containing lifeform (structural) information.

taxonomy

Dataframe containing indigenous status of taxa in bio_df

use_prop_thresh

Numeric. Threshold (proportion) for taxa to include in description. Taxa that occur in more than use_prop_thresh proportion of sites in the cluster will be included in the description.

use_p_val

Numeric 0 to 1. The p-value to use to accept a taxa as an indicator for an ecosystem.

n_ind_max

Maximum number of taxa to list as indicators.

ind_val_iter

Passed to the ... argument of envCluster::make_ind_val_df(), and then into the numitr argument of labdsv::indval().

wt_ht_quant

Numeric. Threshold (quantile) of weighted heights per site that are allowed through to further definition analysis.

sites_sf_quant

Numeric. Threshold (quantile) of counts above which to select structural features for the description.

ht_sf_quant

Numeric. Threshold (quantile) of heights above which to select structural features for the description.

sites_sf_taxa_quant

Numeric. Threshold (quantile) of counts above which a taxa will be used as an example of a structural feature.