Make a (biotic) description for an ecosystem

The description includes: 'obvious' taxa; indicator taxa; taxa present at a 'high' proportion of bins; and structural layers.

make_eco_desc(
  bio_clust_df,
  ind_val_df = NULL,
  context,
  clust_col = "cluster",
  taxa_col = "taxa",
  cov_col = "cover_adj",
  ht_col = "ht",
  lustr,
  str_col = "lifeform",
  indigenous_df = NULL,
  obvious_max_n = 3,
  indicator_p_val = 0.05,
  indicator_max_n = 3,
  common_taxa_prop_thresh = 0.8,
  colour_map = NULL,
  add_structure = TRUE
)

Arguments

bio_clust_df

Dataframe containing all context, all the _cols and taxa data in long format.

ind_val_df

Result from envCluster::make_ind_val_df().

context

Character. Name(s) of column(s) in bio_clust_df that define bins.

clust_col

Character. Name of column in bio_clust_df containing cluster membership.

taxa_col

Character. Name of column in bio_clust_df and/or indigenous_df containing the taxa names.

cov_col

Character. Name of column in bio_clust_df with numeric abundance data (usually 'cover' for plants).

ht_col

Character. Name of column in bio_clust_df with numeric height data.

lustr

Dataframe containing lifeform (structural) information.

str_col

Character name of column in bio_clust_df and lustrcontaining lifeform (or structural) information.

indigenous_df

dataframe containing indigenous status for each unique taxa_col in bio_clust_df

indicator_p_val

Numeric 0 to 1. The p-value to use to accept a taxa as an indicator for an ecosystem.

indicator_max_n

Maximum number of taxa to list as indicators.

common_taxa_prop_thresh

Numeric. Threshold (proportion) for taxa to include in description. Taxa that occur in more than common_taxa_prop_thresh proportion of bins in the cluster will be included in the description.

colour_map

Dataframe mapping any column in result to colour values (in a column called colour).

Value

Tibble of ecosystems descriptions per unique value of clust_col in bio_clust_df

Details

Obvious taxa have large values for per_pres by per_cov_pres by ht.

Indicator taxa requires a result from envCluster::make_ind_val_df() (a wrapper around labdsv::inval()). Choosing indicators uses indicator_p_val and indicator_max_n.

Common taxa are determined as any taxa occurring at more than common_taxa_prop_thresh bins within a cluster.

Obvious and indicator taxa are prevented from also being common taxa.

The structural layer chosen to represent structure for the cluster has the highest value of per_pres by per_cov_pres by ht.